Implement GainData as dataclass
This fixes on the one hand a syntax error in the string formatting
performed within the `__str__` method by removing it and relying on the
method provided by the dataclass.
On the other hand, this change allowed to remove also the `__repr__`,
`__eq__` and `__ne__` methods, which are also provided by the dataclass
wrapper.